Ensure no multiple capital letters appear without using delay
This fixes the issue of capitalizing two letters when SHIFT is active,
but it does not use delay. It instead always updates shift state right
right after a key press, to ensure it is in the right state as soon as
next key is typed. This removes the delay added in #210.